FMS2: open_ASCII_file and open_namelist_file
This patch re-implements the FMS2 implementations of `open_ASCII_file`
and `open_namelist_file` to remove their dependency on FMS1 functions
which have been staged for deletion.

Note that if a file is opened with `mpp_open` but closed with
`close_file_unit`, then it will raise an error in `fms_io_exit`.
This will no longer be an issue after all references to `mpp_open` have
been removed.  But in the meantime, we will need to ensure that all
unit-based `close_file` calls were not opened with `mpp_open`.

There is also a minor patch to `.testing/Makefile` which selects the
framework ("infra") source dependency, rather than hard-set to FMS1.