+SpV_avg optional argument to extract_optics_slice
  Added the optional argument SpV_avg to extract_optics_slice for use along with
an appropriate value for opacity_scale to convert the units of opacity from [Z-1]
to [H-1] in non-Boussinesq mode without making use of the Boussinesq reference
density.  All Boussinesq answers are bitwise identical, but non-Boussinesq
answers will change and become less dependent on the Boussinesq reference density
when this new argument is used.  There is a new optional argument to a publicly
visible subroutine.