+*Non-Boussinesq revision of diabatic_aux
  This commit changes differential_diffuse_TS, diagnoseMLDbyDensityDifference,
diagnoseMLDbyEnergy and applyBoundaryFluxesInOut to make them appropriate for
use in non-Boussinesq mode, and to eliminate dependencies on the Boussinesq
reference density when in that mode.  It also adds a new optional argument to
extract_optics_slice to enable the use of the layer specific volumes to
translate opacities into thickness-based units.

  The specific set of changes include:

 - Add the optional argument SpV_avg to extract_optics_slice and use it along
   with an appropriate value for opacity_scale to optionally convert the units
   of opacity from [Z-1 ~> m-1] to [H-1 ~> m-1 or m2 kg-1] in non-Boussinesq
   mode without making use of the Boussinesq reference density.

 - Use thickness_to_dz and work with internal variables in vertical distances in
   the denominator of diffusive flux calculations in differential_diffuse_T_S,
   diagnoseMLDbyDensityDifference and diagnoseMLDbyEnergy.

 - Refactored diagnoseMLDbyEnergy for probable efficiencies by calling the
   equation of state with contiguous arguments.

 - Use specific volume derivatives to calculate non-Boussinesq mode buoyancy
   fluxes in calculateBuoyancy_Flux1d and applyBoudaryFluxesInOut.

 - Use the inverse of SpV_avg rather than Rho0 in the calculation of the energy
   input used to drive river mixing when in non-Boussinesq mode.

  There are now separate thickness and depth change internal variables in
several places to avoid any dependency on the Boussinesq reference density when
in non-Boussinesq mode.  A total of 8 rescaling factors were eliminated, and in
one place, GV%Rho0 was replaced with GV%H_to_RZ.

  All Boussinesq answers are bitwise identical, but non-Boussinesq answers will
change and become less dependent on the Boussinesq reference density, and there
is a new optional argument to a publicly visible subroutine.